refactor(markdown): move parser into a separate folder
